LayoutTests/platform/ios/fast/text/font-weight-parse-expected.txt

PASS window.getComputedStyle(document.getElementById('test1')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test2')).fontWeight is "1"
PASS window.getComputedStyle(document.getElementById('test3')).fontWeight is "100"
PASS window.getComputedStyle(document.getElementById('test4')).fontWeight is "200"
PASS window.getComputedStyle(document.getElementById('test5')).fontWeight is "300"
PASS window.getComputedStyle(document.getElementById('test6')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test7')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test8')).fontWeight is "500"
PASS window.getComputedStyle(document.getElementById('test9')).fontWeight is "600"
PASS window.getComputedStyle(document.getElementById('test10')).fontWeight is "bold"
PASS window.getComputedStyle(document.getElementById('test11')).fontWeight is "bold"
PASS window.getComputedStyle(document.getElementById('test12')).fontWeight is "800"
PASS window.getComputedStyle(document.getElementById('test13')).fontWeight is "900"
PASS window.getComputedStyle(document.getElementById('test14')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test15')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test16')).fontWeight is "7"
PASS window.getComputedStyle(document.getElementById('test17')).fontWeight is "300"
PASS window.getComputedStyle(document.getElementById('test18')).fontWeight is "200"
PASS window.getComputedStyle(document.getElementById('test19')).fontWeight is "100"
PASS window.getComputedStyle(document.getElementById('test20')).fontWeight is "100"
PASS window.getComputedStyle(document.getElementById('test21')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test22')).fontWeight is "100"
PASS window.getComputedStyle(document.getElementById('test23')).fontWeight is "100"
PASS window.getComputedStyle(document.getElementById('test24')).fontWeight is "123"
PASS window.getComputedStyle(document.getElementById('test25')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test26')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test27')).fontWeight is "normal"
PASS window.getComputedStyle(document.getElementById('test28')).fontWeight is "0"
PASS window.getComputedStyle(document.getElementById('test29')).fontWeight is "1000"
PASS window.getComputedStyle(document.getElementById('test1')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test2')).font is "16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test3')).font is "normal normal 100 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test4')).font is "normal normal 200 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test5')).font is "normal normal 300 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test6')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test7')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test8')).font is "normal normal 500 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test9')).font is "normal normal 600 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test10')).font is "normal normal bold normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test11')).font is "normal normal bold normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test12')).font is "normal normal 800 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test13')).font is "normal normal 900 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test14')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test15')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test16')).font is "16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test17')).font is "normal normal 300 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test18')).font is "normal normal 200 normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test19')).font is "normal normal 100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test20')).font is "normal normal 100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test21')).font is "normal normal normal extra-condensed 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test22')).font is "normal normal 100 normal 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test23')).font is "italic small-caps 100 extra-expanded 48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test24')).font is "48px/49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test25')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test26')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test27')).font is "normal normal normal normal 16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test28')).font is "16px/18px Times"
PASS window.getComputedStyle(document.getElementById('test29')).font is "16px/18px Times"
PASS document.getElementById('test1').style.font is ""
PASS document.getElementById('test19').style.font is "100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test20').style.font is "100 extra-condensed 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test21').style.font is "extra-condensed 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test22').style.font is "100 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test23').style.font is "italic small-caps 100 extra-expanded 48px/49px \"Helvetica Neue\""
PASS document.getElementById('test24').style.font is "italic small-caps 123 extra-expanded 48px/49px \"Helvetica Neue\""
PASS successfullyParsed is true

TEST COMPLETE
